Resolved a few bugs and wrote unit-tests for repositories

+++ src/start/entity.go
@@ -8,8 +8,8 @@ import (
 type Start struct {
 	common.Timestamp
 	ID                   uint                `gorm:"primaryKey"`
-	OptimalDirections    uint64              `gorm:"column:optimaldirection"`
-	SuboptimalDirections uint64              `gorm:"column:suboptimaldirection;size:64"`
+	OptimalDirections    int                 `gorm:"column:optimaldirection"`
+	SuboptimalDirections int                 `gorm:"column:suboptimaldirection;size:64"`
 	StartLocationID      uint                `gorm:"column:startlocationid"`
 	StartLocation        location.Location   `gorm:"foreignkey:StartLocationID;references:id"`
 	Landings             []location.Location `gorm:"many2many:start_landing_locations"`
